Fix managed presets
Presets called managed-requires and managed-build-requires are designed
to contain also artifacts referenced in dependencyManagement sections
of the artifacts already included in the graph. There is a flag in the
respective filters called acceptManaged providing a quick info if these
artifacts should or should not be included.

Based on the flag value filter does not accept managed dependencies if
set to false. If set to true, the decision is passed to the embedded
filter, which was set to not accept managed dependencies regardless of
the flag value.

This commit updates the initialization of the embedded filters to respect
the acceptManaged value.